feat: generate only prompt variant for browser rending
